pyproject.toml
[tool.poetry]
name = "happy"
version = "0.1.0"
description = ""
authors = ["amir_espahbodi <dev.amir.espahbodi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.2"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
sqlalchemy = "^2.0.38"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
